refactorisation styles
