+++ ArgentBank-Frontend/src/components/Button/Button.module.css
@@ -9,18 +9,19 @@
   cursor: pointer;
   border-radius: 5px;
   transition: background-color 0.3s ease;
+  
 }
 
 .button:hover {
   background-color: #009966;
 }
 
-.edit-button {
+/* .editButton {
     border-color: #00bc77;
     background-color: #00bc77;
     color: #fff;
     font-weight: bold;
     padding: 10px;
     border-radius: 5px;
-  }
+  } */
