+++ ArgentBank-Frontend/src/components/EditProfile/EditProfile.module.css
@@ -29,6 +29,7 @@
   font-weight: bold;
   padding: 10px;
   border-radius: 5px;
+  margin-right: 10px;
 }
 
 .editButton:hover {
