+++ ArgentBank-Frontend/src/components/FeatureItem/FeatureItem.module.css
@@ -1,8 +1,8 @@
-.feature-item {
+.featureItem {
     flex: 1;
     padding: 2.5rem;
   }
-  .feature-item-title {
+  .featureItemTitle {
     color: #222;
     font-size: 1.25rem;
     font-weight: bold;
