+++ ArgentBank-Frontend/src/components/Icons/Icon.jsx
@@ -15,28 +15,8 @@ import style from './IconsIndex.module.css';
  */
 
 
-const Icon = ({ size = 'default', type = 'user', className='' }) => { //function Icon({ size = 'default', type = 'user' }) {
-  // Configuration des icônes selon le type et la taille
-  // const iconTypes = {
-  //   // <i className="fa fa-user-circle"></i>,icone user react : <FontAwesomeIcon icon="fa-regular fa-circle-user" /> (icone power de deconnexion : <FontAwesomeIcon icon="fa-light fa-power-off" />), icone paramètre (a utiliser pour EditProfile : <FontAwesomeIcon icon="fa-light fa-gear" /> */}
-  //   default: 'fa fa-user-circle',         // Petit icône pour la nav
-  //   large: 'fa fa-user-circle sign-in-icon' // Grand icône pour le form
-  // }
-//   const iconTypes = {
-//   user: {
-//     default: 'fa fa-user-circle',
-//     large: 'fa fa-user-circle sign-in-icon'
-//   },
-//   settings: {
-//     default: 'fa fa-gear',
-//   },
-//   signout: {
-//     default: 'fa fa-power-off',
-//   }
-// }
-//   return <i className={iconTypes[type][size]}></i>
-// }
-// configuration des icones selonb le type
+const Icon = ({ size = 'default', type = 'user', className }) => { //function Icon({ size = 'default', type = 'user' }) {
+ 
 
 const iconComponents = {
   user: FaUserCircle,
@@ -47,10 +27,10 @@ const iconComponents = {
 const IconComponent = iconComponents[type]
 // defini la taille en fonction du contexte (5rem pour icon user ds loginform et 14px ds nav, cf main.css)
 const iconSize = size === 'large' ? 80 : 16
-const iconClass = size === 'large' ? style['large'] : style['default']
+const iconClass = size === 'large' ? style.iconLarge : style.iconDefault
 
 // retourne l'icone avec la taille et la classe definie dynamiquement
-return <IconComponent size={iconSize} className={`${iconClass}`} />
+return <IconComponent size={iconSize} className={`${iconClass} ${className}`} />
 }
 
 Icon.propTypes = {
