+++ ArgentBank-Frontend/src/components/Icons/IconsIndex.module.css
@@ -3,9 +3,9 @@
     vertical-align: middle;
 }
 
-.icon-large {
+.iconLarge {
     margin-bottom: 1rem;
 }
-.icon-default {
+.iconDefault {
     margin-right: 0.5rem;
 }
